Make it much faster
`importlib.metadata.version()` gives us the version of a package and nothing else that we don't need.

It also returns the versions of editable and file-installed packages. I guess they don't need special casing now... I don't think this will cause any problems but if it does we can fix.